requirements.txt
streamlit
python-dotenv 
google-generativeai 
faiss-cpu
pyPDF2 
langchain 
langchain-huggingface
sentence-transformers
langchain_community